Create a chatbot that speaks to an OpenAI compatible endpoint

This function returns an R6 object that takes care of managing the state associated with the chat; i.e. it records the messages that you send to the server, and the messages that you receive back. If you register a tool (aka an R function), it also takes care of the tool loop.

new_chat_openai(
  system_prompt = NULL,
  base_url = "https://api.openai.com/v1",
  api_key = openai_key(),
  model = "gpt-4o-mini",
  quiet = FALSE
)

Arguments

system_prompt

A system prompt to set the behavior of the assistant.

base_url

The base URL to the endpoint; the default uses ChatGPT.

api_key

The API key to use for authentication. You generally should not supply this directly, but instead set the OPENAI_API_KEY environment variable.

model

The model to use for the chat; defaults to GPT-4o mini.

quiet

If TRUE does not print output as its received.

Methods

Public methods

Chat$new()

Chat$chat()

Chat$register_tool()

Chat$clone()

Method new()

Usage

Chat$new(base_url, model, api_key, system_prompt, quiet = TRUE)

Method chat()

Submit text to the chatbot.

Usage

Chat$chat(text, stream = TRUE)

Arguments

text

The text to send to the chatbot

stream

Whether to stream the response or not.

Method register_tool()

Register a tool (an R function) that the chatbot can use. If the chatbot decides to use the function, elmer will automatically call it and submit the results back.

Usage

Chat$register_tool(name, description, arguments, strict = TRUE)

Arguments

name

The name of the function.

description

A detailed description of what the function does. Generally, the more information that you can provide here, the better.

arguments

A list of arguments that the function accepts. Should be a list of objects created by tool_arg().

strict

Should the argument definition be strictly enforced?

Method clone()

The objects of this class are cloneable with this method.

Usage

Chat$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) { # elmer:::openai_key_exists()
chat <- new_chat_openai()
chat$chat("
  What is the difference between a tibble and a data frame?
  Answer with a bulleted list
")

chat <- new_chat_openai()
chat$register_tool(
  name = "rnorm",
  description = "Drawn numbers from a random normal distribution",
  arguments = list(
    tool_arg(
      "n",
      type = "integer",
      description = "The number of observations. Must be a positive integer."
    ),
    tool_arg(
      "mean",
      type = "number",
      description = "The mean value of the distribution."
    ),
    tool_arg(
      "sd",
      type = "number",
      description = "The standard deviation of the distribution. Must be a non-negative number."
    )
  )
)
chat$chat("
  Give me five numbers from a random normal distribution.
  Briefly explain your work.
")
}