Create a chatbot that speaks to an OpenAI compatible endpoint

This function returns an R6 object that takes care of managing the state associated with the chat; i.e. it records the messages that you send to the server, and the messages that you receive back. If you register a tool (aka an R function), it also takes care of the tool loop.

Usage

new_chat_openai(
  system_prompt = NULL,
  messages = NULL,
  base_url = "https://api.openai.com/v1",
  api_key = openai_key(),
  model = NULL,
  echo = FALSE
)

Arguments

system_prompt

A system prompt to set the behavior of the assistant.

messages

A list of messages to start the chat with (i.e., continuing a previous conversation). If not provided, the conversation begins from scratch. Do not provide non-NULL values for both messages and system_prompt.

Each message in the list should be a named list with at least role (usually system, user, or assistant, but tool is also possible). Normally there is also a content field, which is a string.

base_url

The base URL to the endpoint; the default uses OpenAI.

api_key

The API key to use for authentication. You generally should not supply this directly, but instead set the OPENAI_API_KEY environment variable.

model

The model to use for the chat; set to NULL (the default) to use a reasonable model, currently gpt-4o-mini. We strongly recommend explicitly choosing a model for all but the most casual use.

echo

If TRUE, the chat() method streams the response to stdout by default. (Note that this has no effect on the stream(), chat_async(), and stream_async() methods.)

Value

A promise that resolves to a string (probably Markdown).

Active bindings

system_prompt

The system prompt, if any, as a string.

Methods

Public methods

ChatOpenAI$new()

ChatOpenAI$messages()

ChatOpenAI$chat()

ChatOpenAI$chat_async()

ChatOpenAI$stream()

ChatOpenAI$stream_async()

ChatOpenAI$console()

ChatOpenAI$register_tool()

ChatOpenAI$clone()

Method new()

Usage

ChatOpenAI$new(base_url, model, messages, api_key, echo = FALSE)

Arguments

base_url

The base URL to the endpoint; the default uses ChatGPT.

model

The model to use for the chat; defaults to GPT-4o mini.

messages

An unnamed list of messages to start the chat with (i.e., continuing a previous conversation). If NULL or zero-length list, the conversation begins from scratch.

api_key

The API key to use for authentication. You generally should not supply this directly, but instead set the OPENAI_API_KEY environment variable.

echo

If TRUE, the chat() method streams the response to stdout (while also returning the final response). Note that this has no effect on the stream(), chat_async(), and stream_async() methods.

Method messages()

The messages that have been sent and received so far (optionally starting with the system prompt, if any).

Usage

ChatOpenAI$messages(include_system_prompt = FALSE)

Arguments

include_system_prompt

Whether to include the system prompt in the messages (if any exists).

Method chat()

Submit input to the chatbot, and return the response as a simple string (probably Markdown).

Usage

ChatOpenAI$chat(..., echo = NULL)

Arguments

...

The input to send to the chatbot. Can be strings or images (see content_image_file() and content_image_url().

echo

Whether to emit the response to stdout as it is received. If NULL, then the value of echo set when the chat object was created will be used.

Method chat_async()

Submit input to the chatbot, and receive a promise that resolves with the response all at once.

Usage

ChatOpenAI$chat_async(...)

Arguments

...

The input to send to the chatbot. Can be strings or images.

Method stream()

Submit input to the chatbot, returning streaming results. Returns A coro generator that yields strings. While iterating, the generator will block while waiting for more content from the chatbot.

Usage

ChatOpenAI$stream(...)

Arguments

...

The input to send to the chatbot. Can be strings or images.

Method stream_async()

Submit input to the chatbot, returning asynchronously streaming results. Returns a coro async generator that yields string promises.

Usage

ChatOpenAI$stream_async(...)

Arguments

...

The input to send to the chatbot. Can be strings or images.

Method console()

Enter an interactive chat console. This is a REPL-like interface where you can chat with the assistant in real-time. (Only available in interactive() mode.)

Usage

ChatOpenAI$console()

Method register_tool()

Register a tool (an R function) that the chatbot can use. If the chatbot decides to use the function, elmer will automatically call it and submit the results back.

Usage

ChatOpenAI$register_tool(fun, name, description, arguments, strict = TRUE)

Arguments

fun

The function to be invoked when the tool is called.

name

The name of the function.

description

A detailed description of what the function does. Generally, the more information that you can provide here, the better.

arguments

A named list of arguments that the function accepts. Should be a named list of objects created by tool_arg().

strict

Should the argument definition be strictly enforced?

Method clone()

The objects of this class are cloneable with this method.

Usage

ChatOpenAI$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) { # elmer:::openai_key_exists()
chat <- new_chat_openai()
chat$chat("
  What is the difference between a tibble and a data frame?
  Answer with a bulleted list
")

chat <- new_chat_openai()
chat$register_tool(
  fun = rnorm,
  name = "rnorm",
  description = "Drawn numbers from a random normal distribution",
  arguments = list(
    n = tool_arg(
      type = "integer",
      description = "The number of observations. Must be a positive integer."
    ),
    mean = tool_arg(
      type = "number",
      description = "The mean value of the distribution."
    ),
    sd = tool_arg(
      type = "number",
      description = "The standard deviation of the distribution. Must be a non-negative number."
    )
  )
)
chat$chat("
  Give me five numbers from a random normal distribution.
  Briefly explain your work.
")
}